+++ Firmament/src/utopia/firmament/component/Window.scala
@@ -331,6 +331,7 @@ class Window(protected val wrapped: Either[JDialog, JFrame], container: java.awt
 	component.addWindowListener(WindowStateListener)
 	
 	// Sets up the underlying window
+	component.setLayout(null)
 	component.setContentPane(container)
 	// Some of the functions are only available through the two separate sub-classes
 	wrapped match {
@@ -354,7 +355,10 @@ class Window(protected val wrapped: Either[JDialog, JFrame], container: java.awt
 	_positionPointer.value = Point.of(component.getLocation)
 	_sizePointer.value = Size.of(component.getSize)
 	if (isNotFullScreen)
-		_boundsPointer.onNextChange { _ => centerOnParent() }
+		_boundsPointer.onNextChange { _ =>
+			println("Centers on parent")
+			centerOnParent()
+		}
 	AwtEventThread.async { optimizeBounds(dictateSize = true) }
 	
 	// Registers to update the state when the wrapped window updates
@@ -421,6 +425,7 @@ class Window(protected val wrapped: Either[JDialog, JFrame], container: java.awt
 				updateLayout()
 				content.updateLayout()
 			}
+			component.repaint()
 			DetachmentChoice.continue
 		}
 	}
@@ -600,6 +605,7 @@ class Window(protected val wrapped: Either[JDialog, JFrame], container: java.awt
 	
 	override def position = _positionPointer.value
 	override def position_=(newPosition: Point) = {
+		println(s"Setting position to $newPosition")
 		if (newPosition != position)
 			AwtEventThread.async { component.setLocation(newPosition.toAwtPoint) }
 	}
@@ -619,6 +625,7 @@ class Window(protected val wrapped: Either[JDialog, JFrame], container: java.awt
 	
 	override def bounds: Bounds = _boundsPointer.value
 	override def bounds_=(b: Bounds): Unit = {
+		println(s"Setting bounds to $b")
 		if (b != bounds)
 			AwtEventThread.async { component.setBounds(b.toAwt) }
 	}
@@ -893,14 +900,19 @@ class Window(protected val wrapped: Either[JDialog, JFrame], container: java.awt
 	
 	private object WindowComponentStateListener extends ComponentAdapter
 	{
+		
 		override def componentShown(e: ComponentEvent) = _visibleFlag.set()
 		override def componentHidden(e: ComponentEvent) = _visibleFlag.reset()
 		
-		override def componentMoved(e: ComponentEvent) = _positionPointer.value = Point.of(component.getLocation)
+		override def componentMoved(e: ComponentEvent) = {
+			println("Window moved")
+			_positionPointer.value = Point.of(component.getLocation)
+		}
 		// NB: This will not limit user's ability to resize window beyond minimum and maximum
 		override def componentResized(e: ComponentEvent) = {
 			val newSize = Size.of(component.getSize)
 			_sizePointer.value = newSize
+			
 			// Updates content layout (only while visible)
 			if (isFullyVisible) {
 				updateLayout()